+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # `nonempty-collections`
 
+## 1.1.0 (2025-10-16)
+
+#### Added
+
+- The `Singleton` trait.
+- `NonEmptyIterator::unzip`.
+- `Extend` impls for `NESet`, `NEMap`, and `NEIndexMap`.
+
 ## 1.0.1 (2025-09-06)
 
 #### Added
 
-- Various `AsRef` imps.
+- Various `AsRef` impls.
 
 #### Fixed